backend/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.8
django-filter==2.1.0
django-filters==0.2.1
django-rest-swagger==2.2.0
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mysqlclient==1.3.12
openapi-codec==1.3.2
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
gevent==1.4.0
